[core] Clean up extension code a bit.
